# utils/swaggerApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "test API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://yurgo.fun"
    }
  ],
  "tags": [
    {
      "name": "Contacts API",
      "description": "API to work with contacts"
    }
  ],
  "paths": {
    "/api/contacts/": {
      "get": {
        "summary": "Retrieve a list of contacts. Required authorization by Bearer token",
        "tags": [
          "Contacts API"
        ],
        "description": "parameter id just for test",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pagination": {
                      "type": "object",
                      "properties":{
                        "totalContacts":{
                          "type":"integer",
                          "example": "98"
                        },
                        "page": {
                          "type": "integer",
                          "example": "1"
                        },
                        "perPage":{
                          "type": "integer",
                          "example": "5"
                        }
                      }
                    },
                    "contacts": {
                      "type":"array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id":{
                            "type": "string",
                            "example": "642c87122969c63fa569e028"
                          },
                          "name":{
                            "type": "string",
                            "example": "Michael Douglas"
                          },
                          "email":{
                            "type": "string",
                            "example": "email@email.com" 
                          },
                          "phone":{
                            "type": "string",
                            "example": "+380670000001" 
                          },
                          "favorite":{
                            "type": "boolean",
                            "example": "false" 
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "returns simple answer from post",
        "tags": [
          "Contacts API"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Info"
              },
              "example": {
                "country": "Russia",
                "city": "Spb"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "simple parameter",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "description": "parameter id just for test",
        "responses": {
          "204": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Info": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}